% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/mlasiagitstat.R
\name{mlasiagitstat}
\alias{mlasiagitstat}
\title{Plot github stats for Media Lab Asia}
\usage{
mlasiagitstat(id = "poisoncreed", type = c("users", "orgs"), max = 20)
}
\arguments{
\item{id}{name of the github user or organization}

\item{type}{either "users" or "orgs"}

\item{max}{maximum number of repositories to plot}
}
\description{
This function plots the total number of forks and followers for all repositories of a certain user or organization.
}
\author{
Vishal Virmani
}

